requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
chex==0.1.6
click==8.1.3
cmake==3.26.0
comm==0.1.3
contourpy==1.0.7
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
diffrax==0.1.2
dm-haiku==0.0.9
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.0
entrypoints==0.4
equinox==0.10.1
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
filelock==3.10.0
flatbuffers==23.3.3
fonttools==4.39.2
FyeldGenerator==0.1.7
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.25.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.4
jax==0.4.6
jaxlib==0.4.6+cuda11.cudnn82
jaxtyping==0.2.14
jedi==0.18.2
Jinja2==3.1.2
jmp==0.0.4
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-server==1.23.4
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
lit==15.0.7
lpips==0.1.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==0.8.4
moviepy==1.0.3
mpmath==1.3.0
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
phiflow==2.4.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.18.0
ply==3.11
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-version==0.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.17.0
setproctitle==1.3.2
setuptools==65.6.3
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
tifffile==2023.3.15
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typeguard==3.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.14.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.15.0
y-py==0.5.9
yarg==0.1.9
ypy-websocket==0.8.2
zipp==3.15.0
